exception

+++ src/main/java/com/stackroute/muzixtrack/service/TrackServiceImpl.java
@@ -1,6 +1,8 @@
 package com.stackroute.muzixtrack.service;
 
 import com.stackroute.muzixtrack.domain.Track;
+import com.stackroute.muzixtrack.exceptions.TrackAlreadyExistsException;
+import com.stackroute.muzixtrack.exceptions.TrackNotFoundException;
 import com.stackroute.muzixtrack.repository.TrackRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -19,39 +21,70 @@ class TrackServiceImpl implements TrackService {
   }
 
   @Override
-  public Track save(Track track) {
+  public Track save(Track track) throws TrackAlreadyExistsException {
+    if (trackRepository.existsById(track.getId())) {
+      throw new TrackAlreadyExistsException(("user already exists"));
+    }
     Track savedTrack = trackRepository.save(track);
+    if (savedTrack == null) {
+      throw new TrackAlreadyExistsException("user already exists");
+    }
     return savedTrack;
   }
 
   @Override
-  public Track getById(int id) {
+  public Track getById(int id) throws TrackNotFoundException {
+    if (!trackRepository.existsById(id)) {
+      throw new TrackNotFoundException("not found");
+    }
     Track retriveTrack = trackRepository.findById(id).get();
     return retriveTrack;
   }
 
   @Override
-  public List<Track> getAllTracks() {
-    return trackRepository.findAll();
+  public List<Track> getAllTracks() throws Exception {
+    if (trackRepository.findAll().isEmpty()) {
+      throw new Exception("Internal server error");
+    } else {
+      List<Track> allTracks = trackRepository.findAll();
+      return allTracks;
+    }
   }
 
   @Override
-  public Track deleteById(int id) {
-    Optional<Track> optionalTrack = trackRepository.findById(id);
-    trackRepository.deleteById(id);
-    return optionalTrack.get();
+  //delete a particular  part of a track by its id
+  public Track deleteById(int id) throws TrackNotFoundException {
+    //  delete the track
+    if (trackRepository.existsById(id)) {
+      Track retrivedTrack = trackRepository.findById(id).get();
+      trackRepository.deleteById(id);
+      return retrivedTrack;
+    } else {
+      throw new TrackNotFoundException("Track not found");
+    }
   }
 
   @Override
-  public Track updateTrackbyId(int id, Track track) {
-    Optional<Track> optionalTrack = trackRepository.findById(id);
-    Track updateTrack = trackRepository.save(track);
-    return updateTrack;
+  //update a particular  part of a track by its id
+  public Track updateTrackById(int id,Track track) throws TrackNotFoundException {
+//  update the track
+    if (trackRepository.existsById(id)) {
+      Track updateTrackById = trackRepository.findById(id).get();
+      return updateTrackById;
+    } else {
+      throw new TrackNotFoundException("track not found");
+    }
   }
+
 
   @Override
-  public List<Track> trackByName(String name) {
-    return trackRepository.trackByName(name);
+  public List<Track> trackByName(String name) throws Exception {
+    if (trackRepository.trackByName(name).isEmpty()) {
+      throw new Exception("no track with this name");
+    } else {
+      List<Track> allTracksByName = trackRepository.trackByName(name);
+      return allTracksByName;
+    }
   }
-
 }
+
